Update kernel 2.6.8 to 2.6.12
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@1115 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/mtd/maps/arctic-mtd.c
@@ -98,7 +98,7 @@ init_arctic_mtd(void)
 {
 	printk("%s: 0x%08x at 0x%08x\n", NAME, SIZE, PADDR);
 
-	arctic_mtd_map.virt = (unsigned long) ioremap(PADDR, SIZE);
+	arctic_mtd_map.virt = ioremap(PADDR, SIZE);
 
 	if (!arctic_mtd_map.virt) {
 		printk("%s: failed to ioremap 0x%x\n", NAME, PADDR);
