+++ linux-2.6.x/drivers/mtd/maps/fortunet.c
@@ -209,8 +209,8 @@ int __init init_fortunet(void)
 
 			map_regions[ix].map_info.phys =	map_regions[ix].window_addr_physical,
 
-			map_regions[ix].map_info.virt =
-				(int)ioremap_nocache(
+			map_regions[ix].map_info.virt = 
+				ioremap_nocache(
 				map_regions[ix].window_addr_physical,
 				map_regions[ix].map_info.size);
 			if(!map_regions[ix].map_info.virt)
