+++ linux-2.6.x/drivers/mtd/maps/pb1550-flash.c
@@ -178,8 +178,7 @@ int __init pb1550_mtd_init(void)
 	 */
 	printk(KERN_NOTICE "Pb1550 flash: probing %d-bit flash bus\n", 
 			pb1550_map.bankwidth*8);
-	pb1550_map.virt = 
-		(unsigned long)ioremap(window_addr, window_size);
+	pb1550_map.virt = ioremap(window_addr, window_size);
 	mymtd = do_map_probe("cfi_probe", &pb1550_map);
 	if (!mymtd) return -ENXIO;
 	mymtd->owner = THIS_MODULE;
